+++ R/bayesfactor_models.R
@@ -68,7 +68,7 @@
 #'   non-log Bayes factors; see examples), that prints nicely.
 #'   \cr\cr
 #'   For `as.matrix()` a square matrix of (log) Bayes factors, with rows as
-#'   numerators and columns as denominators.
+#'   denominators and columns as numerators.
 #'
 #' @examplesIf require("lme4") && require("BayesFactor") && require("rstanarm") && require("brms")
 #' # With lm objects: